requirements.txt
asgiref==3.8.1
Django==3.2.14   
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
install==1.3.5
isort==5.13.2
packaging==24.1
pillow==10.3.0  
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
